kafka

+++ OCR/api/requestToRedis/schemas.py
@@ -4,8 +4,8 @@ from typing import List
 class Product(BaseModel):
     name: str
     numberServings: int
-    price: float
-    total: float
+    price: int
+    total: int
     
 class SCache(BaseModel):
     numberClients: int
